% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coordinates_to_wkt_square_polygon}
\alias{coordinates_to_wkt_square_polygon}
\title{Convert Coordinates to WKT Square Polygon}
\usage{
coordinates_to_wkt_square_polygon(lat, lon, buffer_distance)
}
\arguments{
\item{lat}{Numeric. Latitude of the center of the square.}

\item{lon}{Numeric. Longitude of the center of the square.}

\item{buffer_distance}{Numeric. The distance from the center to each side of the square.}
}
\value{
A character string representing the square polygon in WKT format.
}
\description{
This function generates a square polygon in Well-Known Text (WKT) format
centered on the provided coordinates and with the specified buffer distance.
}
\examples{
coordinates_to_wkt_square_polygon(lat = 45.0, lon = -93.0, buffer_distance = 0.01)
}
