% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_directory}
\alias{create_directory}
\title{Create a directory}
\usage{
create_directory(dir_name)
}
\arguments{
\item{dir_name}{A character string of the directory name to create.}
}
\description{
This function creates a directory if it doesn't already exist.
}
