% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download.file.int}
\alias{download.file.int}
\title{Download a file}
\usage{
download.file.int(url, destfile)
}
\arguments{
\item{url}{A character string of the URL to download from.}

\item{destfile}{A character string of the destination file path.}
}
\description{
An internal function to download a file from a URL.
}
\keyword{internal}
