% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_search.R
\name{download.specimen.image}
\alias{download.specimen.image}
\title{Download specimen images}
\usage{
download.specimen.image(
  metadata,
  resize = c(""),
  dir_name = "my_virtual_collection"
)
}
\arguments{
\item{metadata}{A data frame containing specimen metadata, as returned by search.specimen.metadata().}

\item{resize}{A character vector specifying resize options (currently not used in the function).}

\item{dir_name}{A character string specifying the directory to save the downloaded images.}
}
\description{
This function downloads specimen images based on the provided metadata and optionally resizes them.
}
