% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_with_retry}
\alias{download_with_retry}
\title{Download a File with Retry Mechanism}
\usage{
download_with_retry(url, destfile, max_attempts)
}
\arguments{
\item{url}{Character. The URL from which to download the file.}

\item{destfile}{Character. The destination file path where the downloaded file will be saved.}

\item{max_attempts}{Integer. The maximum number of download attempts before giving up.}
}
\value{
Logical. TRUE if the download succeeded, FALSE otherwise.
}
\description{
This function attempts to download a file from a URL multiple times until
it succeeds or the maximum number of attempts is reached.
}
