% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_specimens_by_country}
\alias{plot_specimens_by_country}
\title{Plot the Number of Specimens by Country}
\usage{
plot_specimens_by_country(metadata)
}
\arguments{
\item{metadata}{A data frame containing at least the column \code{country},
which lists the country associated with each specimen.}
}
\value{
A bar plot showing the number of specimens for each country.
}
\description{
This function takes metadata containing specimen information and plots the
number of specimens for each country as a bar plot.
}
