% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_specimens_by_institution}
\alias{plot_specimens_by_institution}
\title{Plot the Number of Specimens by Institution}
\usage{
plot_specimens_by_institution(metadata)
}
\arguments{
\item{metadata}{A data frame containing at least the column \code{institutionCode},
which lists the institution associated with each specimen.}
}
\value{
A bar plot showing the number of specimens for each institution.
}
\description{
This function takes metadata containing specimen information and plots the
number of specimens for each institution as a bar plot.
}
