% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{resize.image}
\alias{resize.image}
\title{Resize an image}
\usage{
resize.image(file_name, current_megapixels, max_megapixels)
}
\arguments{
\item{file_name}{A character string of the file path of the image to resize.}

\item{current_megapixels}{A numeric value of the current megapixels.}

\item{max_megapixels}{A numeric value of the maximum desired megapixels.}
}
\value{
A magick image object of the resized image.
}
\description{
This function resizes an image to fall within a specified megapixel range.
}
