% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_search.R
\name{search.specimen.metadata}
\alias{search.specimen.metadata}
\title{Search for specimen metadata}
\usage{
search.specimen.metadata(species_name, ...)
}
\arguments{
\item{species_name}{A character string of the scientific name of the species to search for.}

\item{...}{Additional arguments passed to occ_search function.}
}
\value{
A data frame containing metadata for the found specimens, including media URLs and licenses.
}
\description{
This function searches GBIF for records with images of preserved specimens for a given species.
}
