% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_search.R
\name{download_specimen_images}
\alias{download_specimen_images}
\title{Download specimen images}
\usage{
download_specimen_images(
  metadata,
  dir_name = "my_virtual_collection",
  resize = NULL
)
}
\arguments{
\item{metadata}{A data frame containing specimen metadata, as returned by search.specimen.metadata().}

\item{dir_name}{A character string specifying the directory to save the downloaded images.}

\item{resize}{A character vector specifying resize options (currently not used in the function).}
}
\description{
This function downloads specimen images based on the provided metadata and optionally resizes them.
}
