% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_search.R
\name{search_specimen_metadata}
\alias{search_specimen_metadata}
\title{Search for Specimen Metadata}
\usage{
search_specimen_metadata(
  taxon_name = NULL,
  coordinates = NULL,
  buffer_distance = NULL,
  limit = 500,
  verbose = TRUE,
  search_type = "herbarium",
  user = NULL,
  pwd = NULL,
  email = NULL,
  ...
)
}
\arguments{
\item{taxon_name}{A character string representing the scientific name of the taxon to search for.}

\item{coordinates}{A numeric vector representing a coordinate point, passed as latitude and longitude (e.g., \code{c(42, -85)}), to be the centroid of the polygon where the search will be performed.}

\item{buffer_distance}{Numeric. How far from the coordinates to search for specimens.}

\item{limit}{Numeric. The maximum number of records to search for on GBIF. Defaults to 500.}

\item{verbose}{Should the function print the number of records found with media data on console? Default is TRUE.}

\item{search_type}{The default is "herbarium" for herbarium specimens. The other option is "cs" for citizen science images (mainly living plants from imported from iNaturalist).}

\item{user}{GBIF username. If provided, the function will send a formal download request to GBIF and add a column in the output with the proper DOI citation (RECOMMENDED FOR PUBLICATIONS).}

\item{pwd}{GBIF password. If provided, the function will send a formal download request to GBIF and add a column in the output with the proper DOI citation (RECOMMENDED FOR PUBLICATIONS).}

\item{email}{GBIF email. If provided, the function will send a formal download request to GBIF and add a column in the output with the proper DOI citation (RECOMMENDED FOR PUBLICATIONS).}

\item{...}{Additional arguments passed to the \code{occ_search} function.}
}
\value{
A data.frame containing metadata for the found specimens, including media URLs and licenses.
}
\description{
This function searches GBIF for records with images of preserved specimens for a given species.
}
\examples{
\donttest{
metadata <- search_specimen_metadata(taxon_name = "Vaccinium", coordinates = c(42, -85))
}
}
